feat: update person detection to GA
